SISRP-45434 - Post-3/3 React Refactor - move error/spinner logic from Widget to WidgetBody
