+++ src/react/components/base/widget/Widget.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Spinner from '../Spinner';
 import WidgetBody from './WidgetBody';
 import WidgetHeader from './WidgetHeader';
 
@@ -9,50 +8,21 @@ import '../../../stylesheets/box_model.scss';
 import '../../../stylesheets/print.scss';
 import '../../../stylesheets/widgets.scss';
 
-const renderErrorMessage = (errorMessage) => {
-  let messageComponent;
-  if (typeof errorMessage === 'string' || errorMessage instanceof String) {
-    messageComponent = (<p className="cc-react--no-margin">{errorMessage}</p>);
-  } else {
-    messageComponent = errorMessage;
-  }
-  return (
-    <div className="cc-react-widget--padding">
-      {messageComponent}
-    </div>
-  );
-};
-
-const renderWidgetBody = (widgetConfig, children) => {
-  if (widgetConfig.errored) {
-    return renderErrorMessage(widgetConfig.errorMessage);
-  } else if (widgetConfig.isLoading) {
-    return <Spinner isLoadingMessage={widgetConfig.isLoadingMessage} />;
-  } else {
-    return (
-      <WidgetBody padding={widgetConfig.padding}>
-        {children}
-      </WidgetBody>
-    );
-  }
-};
-
-const renderWidget = (widgetConfig, children) => {
+const Widget = (props) => {
+  const {config: widgetConfig, children} = props;
   if (widgetConfig.visible) {
     return (
       // class "cc-widget" is necessary for now for 2-column and 3-column margins. See _widgets.scss
       <div className="cc-react-widget cc-widget">
         <WidgetHeader title={widgetConfig.title} />
-        {renderWidgetBody(widgetConfig, children)}
+        <WidgetBody widgetConfig={widgetConfig}>
+          {children}
+        </WidgetBody>
       </div>
     );
   } else {
     return null;
   }
-};
-
-const Widget = (props) => {
-  return renderWidget(props.config, props.children);
 };
 
 Widget.defaultProps = {
